Solved HSB detection problem
Before the camera did not want to render but the console log said it was displaying. The issue was with the mouse.Pressed function for detecting HSB values. Removing that line of code and moving it to the draw function and changing it to a if else statement, now it works and detects HSB values. Used ChatGPT to know where to move the code for it to work.